# openapi.yaml
openapi: 3.0.0
info:
  title: YouTube Multi API
  description: API for downloading YouTube videos as MP3/MP4 and getting AI-processed transcripts
  version: 1.0.0
  contact:
    name: BaySercan
    url: https://github.com/BaySercan/youtube-multi-api
servers:
  - url: http://localhost:3500
    description: Local development server
  - url: https://youtube-multi-api.onrender.com
    description: Production server (Render.com)
    
security:
  - RapidAPI: []
    
components:
  securitySchemes:
    RapidAPI:
      type: apiKey
      name: X-RapidAPI-Proxy-Secret
      in: header
      description: RapidAPI authentication token
paths:
  /ping:
    get:
      summary: Health check endpoint
      description: No authentication required
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2025-06-01T16:00:00.000Z"
                  version:
                    type: string
                    example: "1.0.0"
  /info:
    get:
      summary: Get video metadata
      description: Retrieve metadata about a YouTube video
      security:
        - RapidAPI: []
      parameters:
        - in: query
          name: url
          required: true
          schema:
            type: string
          description: YouTube video URL
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                  thumbnail:
                    type: string
                  video_id:
                    type: string
                  channel_id:
                    type: string
                  channel_name:
                    type: string
                  post_date:
                    type: string
              example:
                title: "Example Video Title"
                thumbnail: "https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg"
                video_id: "dQw4w9WgXcQ"
                channel_id: "UC12345678"
                channel_name: "Example Channel"
                post_date: "20230101"
        '400':
          description: Invalid request
  /mp3:
    get:
      summary: Download video as MP3
      description: Download a YouTube video as an MP3 audio file
      security:
        - RapidAPI: []
      parameters:
        - in: query
          name: url
          required: true
          schema:
            type: string
          description: YouTube video URL
      responses:
        '200':
          description: MP3 file download
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid request
  /mp4:
    get:
      summary: Download video as MP4
      description: Download a YouTube video as an MP4 video file
      security:
        - RapidAPI: []
      parameters:
        - in: query
          name: url
          required: true
          schema:
            type: string
          description: YouTube video URL
      responses:
        '200':
          description: MP4 file download
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid request
  /transcript:
    get:
      summary: Get video transcript
      description: Retrieve AI-processed video transcript
      security:
        - RapidAPI: []
      parameters:
        - in: query
          name: url
          required: true
          schema:
            type: string
          description: YouTube video URL
        - in: query
          name: lang
          schema:
            type: string
            default: tr
          description: Language code (e.g., en, tr, es)
        - in: query
          name: skipAI
          schema:
            type: boolean
            default: false
          description: Skip AI processing
        - in: query
          name: useDeepSeek
          schema:
            type: boolean
            default: true
          description: Use DeepSeek model for processing
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  title:
                    type: string
                  language:
                    type: string
                  transcript:
                    type: string
                  ai_notes:
                    type: string
                  isProcessed:
                    type: boolean
                  processor:
                    type: string
                  video_id:
                    type: string
                  channel_id:
                    type: string
                  channel_name:
                    type: string
                  post_date:
                    type: string
              example:
                success: true
                title: "Example Video Title"
                language: "tr"
                transcript: "This is the cleaned transcript text..."
                ai_notes: "Processing notes from AI model"
                isProcessed: true
                processor: "deepseek"
                video_id: "dQw4w9WgXcQ"
                channel_id: "UC12345678"
                channel_name: "Example Channel"
                post_date: "20230101"
        '400':
          description: Invalid request or no subtitles available
          
  /progress/{id}:
    get:
      summary: Get processing status
      description: Retrieve the current status of a processing job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Processing job ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                  progress:
                    type: number
                  video_id:
                    type: string
                  video_title:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
                  lastUpdated:
                    type: string
                    format: date-time
        '404':
          description: Processing ID not found
          
  /result/{id}:
    get:
      summary: Get processing result
      description: Retrieve the final result of a completed processing job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Processing job ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  title:
                    type: string
                  language:
                    type: string
                  transcript:
                    type: string
                  ai_notes:
                    type: string
                  isProcessed:
                    type: boolean
                  processor:
                    type: string
                  video_id:
                    type: string
                  channel_id:
                    type: string
                  channel_name:
                    type: string
                  post_date:
                    type: string
        '202':
          description: Processing not complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: string
                  progress:
                    type: number
        '404':
          description: Processing ID not found